added in the little rolly ball for controlling the filter, now got a full monosample path. everything seems to be hunky dory for now.

+++ src/gridController.mm
@@ -90,7 +90,7 @@
 			if(steps[j] == i){
 				ofFill();
 			}else{
-				if(fmod(player.tick, 8)!=j){
+				if(fmod((double)player.tick, (double)8)!=j){
 					ofSetColor(0, 0, 0);
 				}
 				ofNoFill();
